+++ src/components/Navigation/ModeDropdown.js
@@ -1,20 +1,24 @@
 import React, {PropTypes} from 'react';
+import {connect} from 'react-redux';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import styles from './ModeDropdown.scss';
 import MenuItem from 'react-bootstrap/lib/MenuItem';
 import DropdownButton from 'react-bootstrap/lib/DropdownButton';
+import {getTranslator} from '../../selectors/translate';
+import {setModeStudent, setModeTeacher} from '../../action_creators';
 
-const modes = ['student', 'teacher'];
-const ModeDropdown = ({t, setModeStudent, setModeTeacher, mode}) => {
+const ModeDropdown = ({t, setModeStudent, setModeTeacher, isStudentMode}) => {
+  const modes = ['student', 'teacher'];
   const texts = {'student': t('general.student'), 'teacher': t('general.teacher')};
-  const setMode = mode => mode === 'student' ? setModeStudent() : setModeTeacher();
+  const mode = isStudentMode ? 'student' : 'teacher';
+  const setMode = isStudent => isStudent ? setModeTeacher() : setModeStudent();
   return <div className={styles.gadgetContainer}>
     <DropdownButton id='mode-dropdown'
                     noCaret
                     pullRight
                     bsStyle={mode}
                     title={t('navbar.mode') + ': ' + texts[mode]}
-                    onSelect={setMode}>
+                    onSelect={() => setMode(isStudentMode)}>
       {
         modes.map(k =>
           <MenuItem key={k} eventKey={k} active={mode === k}>{texts[k]}</MenuItem>
@@ -25,15 +29,26 @@ const ModeDropdown = ({t, setModeStudent, setModeTeacher, mode}) => {
 };
 
 ModeDropdown.propTypes = {
-  // ownProps
-  mode: PropTypes.oneOf(modes).isRequired,
-
   //mapStateToProps
   t: PropTypes.func.isRequired,
+  isStudentMode: PropTypes.bool.isRequired,
 
   // mapDispatchToProps
   setModeStudent: PropTypes.func.isRequired,
   setModeTeacher: PropTypes.func.isRequired
 };
 
-export default (withStyles(styles)(ModeDropdown));
+const mapStateToProps = (state) => ({
+  isStudentMode: state.isStudentMode,
+  t: getTranslator(state)
+});
+
+const mapDispatchToProps = {
+  setModeStudent,
+  setModeTeacher
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(withStyles(styles)(ModeDropdown));
