+++ src/components/Navigation/NavBar.scss
@@ -155,8 +155,3 @@ $widget-height: 50px;
   float: right;
   padding: 4px 0;
 }
-
-.gadgetContainer {
-  display: inline-block;
-  padding: 3px 5px;
-}
